flux/ai/agent-contracts/src/requirements.txt
# === Smart Contract Analysis ===
slither-analyzer>=0.10.0
py-solc-x>=2.0.0

# === Blockchain Interaction ===
web3>=7.0.0
eth-abi>=5.0.0

# === LLM Integration ===
anthropic>=0.30.0
httpx>=0.27.0

# === CloudEvents & Messaging ===
cloudevents>=1.12.0  # Latest secure version
pika>=1.3.0
aio-pika>=9.0.0

# === Web Framework ===
fastapi>=0.115.6  # Fixed CVE-2024-24762
uvicorn>=0.32.1
pydantic>=2.12.5  # Latest stable
python-dotenv>=1.0.0

# === Storage & Caching ===
boto3>=1.34.0
redis>=5.0.0

# === Observability ===
opentelemetry-api>=1.30.0
opentelemetry-sdk>=1.30.0
prometheus-client>=0.21.0
structlog>=24.4.0

# === Utilities ===
tenacity>=8.3.0
pyyaml>=6.0.0
jinja2>=3.1.0

# === Agent Optimization (Queueing, Game, Control Theory) ===
# Installed from local shared-lib during Docker build
# -e /app/shared-lib
